CLI tutorial (markdown)
This adds a basic CLI tutorial. The goal is to walk the user through the
process of setting up, running, and gathering results using the CLI.

Audience is people for whom this is their first experience with OpenFE,
and they're not super into Python. They already know the science,
however, and may know some of the tools in this space (LOMAP, etc).